// frontend/src/types/types2.ts
export interface BrokerCommissionReportFilters {
  broker_id?: number;
  start_date?: string;
  end_date?: string;
  filter_type?: string;
}
// ...
export interface TaxReportFilters {
  start_date?: string;
  end_date?: string;
  filter_type?: string;
}
// ...
export interface AdminTaxReportFilters {
  user_id?: number;
  start_date?: string;
  end_date?: string;
  filter_type?: string;
}